Verbatim copy of my previous TT05 design
This is a copy of [1], except for the changes in the layout of files and
directories.

Another difference is that it's missing this change that used to go to
`src/config.tcl`, and now is supposed to go to `src/config.json`. I'm
not sure at this point I will need it, so I'm excluding it.

    -set ::env(PL_TARGET_DENSITY) 0.6
    +set ::env(PL_TARGET_DENSITY) 0.65

[1]: https://github.com/rtfb/tt05-collatz