Reset orbit_len when entering compute mode
When the ASIC starts working on a new input, it should reset the orbit
length, otherwise we can't feed it another input right after reading out
an output of a completed orbit.

This bug could have been worked around in a couple ways:

1. Reset the design. `orbit_len` is initialized to zero in the reset
   clause, so this helped. However, it wastes time in between the
   inputs.

2. Keep track of the old value of orbit_len and subtract it in the
   driver. This should work, but sooner or later `orbit_len` would
   overflow and it would be cumbersome to handle this case.

This commit fixes this problem without the drawbacks of the workarounds.