Reset iterator after the orbit gets computed
This is a minor optimization/clarification. Set the iter register to all
zeros upon reset and after each orbit. This allows skipping the zero
bytes when setting the input number.

Realistically, I don't expect much speed gain from skipping an
occasional IO operation, but I was bitten by forgetting to set the upper
bytes to zeros when testing, so at least this makes it a bit more tidy
from the software perspective.